+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 * Reload lua scripts when modified by user (#519)
 * Debug errors show file and line
 * Add markers to frames in input editor (#548)
+* Add an option to pause on markers
 
 ### Changed